+++ src/pagesRedesign/components/HeaderRedesign.tsx
