+++ src/pagesRedesign/components/SideBarRedesign.tsx
